Renamed "StagedData" to "Cleaned/CleanData" to better align with what the code is actually doing.
